requirements.txt
tornado==5.0.2
requests[security]==2.11.1
gcloud==0.18.3
google-cloud-speech==0.33.0
grpcio==1.0.0
mock==2.0.0
